Add new view for characteristics to JSON
It's a pain to have to query the EAV design pattern to get
characteristic values, so we make it somewhat easier by reducing all
assigned characteristics into a single record where the characteristic
assignments are presented as JSON.

In future we may want to take this same approach and make views that
join to the target table type assumed by the characteristic assignment
so that you can query, for example, for an item and get the item data
and all characteristic data in a single, simple query.  But this is a
start.