Add root cause exception reporting option
One of the problems with the way we've been handling exceptions is that
we collect a wonderful amount of information, but often times the actual
problem is buried at the bottom of a stack trace... which isn't even
visible to the user.  This option rectifies that by searching for the
now conventional "thrownError" reference in the payload.  We trace that
down until we run out and that (we presume) is the root cause.  If the
new config is true, we display the root cause messages which should be
much more useful.

We also do a minor refactoring of the isDebugging flag to make it a
MuseUtils public property... we avoid the "this" issues and we can set
it interactively during debugging which can be helpful.