Add createNumericLineEdit function
The previous method of getting the numeric line function extensions
was to create the widget and then send it to numericLineEdit and rely
on side effects of that call.  The new function creates the widget for
you, runs numericLineEdit, and then returns the new widget.

The old method is not deprecated for backwards compatibility and is in
fact used by the new funciton.