Add event hook framework save event functionality
Here we deal explicitly with the "save" event hooks.  We could have done
this in a more generic function as to deal with all possible events that
we might want to hook, but other than save, the other events (i.e.
close) are much more speculative.  If the need arises we'll revise, but
this is a good enough start that we can at least prove out the method.