Refactor save event hook framework code
We modularize the framework in an init function which returns an
instance of an object containing the rest of the functionality.  This
-should- allow multiple instances of the framework if it's appropriate
to have multiple instances; I'm pretty sure the init function will
return a new instance each time it's run, but naturally I could be wrong
as this is a little outside the norm of what I usually do (and, you
know, JavaScript).