Add optional logging with pseudo-severities
This is something of a departure from the original tickets.  After some
thinking we really needed a way to say whether or not to log so it's
implemented as a true/false conditional check (default false).  We did
add some constants which are set either true or false so we could have
a single place to turn up/turn down the log levels.  At the end of the
day, this seemed the best option when trade-offs between performance,
storage, and log useful were all considered.