Add null handling to numericLineEdit
Prior to this feature, setFormattedText had to be a non-null value with
some numbers embedded somewhere in the text such that numbro could parse
a numeric value out of it.  However, this isn't really correct as there
are times, even in a UI, that a numeric field is truly undefined and 0
would convey too much meaning (or be indistingishable from an
intentional 0.)

We add the ability to set and retrieve null values, though we make some
concessions to null returns by not doing it by default.