Add new exception display dialog box
The prior QMessageBox way of displaying our exceptions was simple and
performant, but didn't allow users to copy and paste errors into other
applications and didn't provide very rich information.  Out approach
where is a bit more heavy handed than we'd like (full UI and backing
script), and not so visually appealing as the older way, but much more
functional.

If this becomes a performance issue (loading the form UI/script from the
database) we can make a window inside the regular exception script, but
for now I think this approach is OK.