Add Item Availabilty Workbench utility script
This script forces the embedded Item Master form's scripts to load and
initializes them when a new workbench item is selected.  This forces a
load on an otherwise lazily loaded and non-initializing (via "set")
process that the out of box application does by default.

(This is the hotfix edition)

Addresses:
https://gitlab.com/musesystems-opensource/musextputils/issues/14