Fix overly restrictive validation; mode enum func
In the function which returns the textual value for a given mode enum value,
the public API function stopped any values less than 0 passing.  This was too
restrictive as the function already had a proper return value for out-of-bounds
values like -1 (which can be passed from the C++ Qt code).

There is a strong argument that this is a change to a public API that should
cause a major version bump, but in this case I think we're ultimately justified
in not doing that.  This change should be backward compatible based on the
purpose of the function (callers should already be able to deal with the change
in behavior).  Moreover, we don't regularly get error reports which would have
been the result of the old behavior which means that in most cases where a
caller may care, we already get expected, valid enum values.  On this basis and
the work required to update all the existing version 3 stuff to version 4...
this is really just a small, backward compatible in practice, bugfix.

Related to: https://gitlab.com/musesystems-rwm/support/issues/22