Start Muse Architecture feature set
Primarily this has to do with being able to have the programming paradigm that
was developed in musesystems-opensource/musesuperchar for all custom xtuple
development.  While superchars address database and layout issues, we'll really
just focus on providing a standard way to do more reactive programming with less
Qt/xTuple boilerplate.

This is the start of that to start showing some intentions related to how this
will work.  Later, we will have the superchar module itself become more of a
musearch extension rather than continue to generate all of the artifacts that it
does today... this may allow us to make it a more general purpose form builder.