Fix inability to get view based data models
Getting data models required not just a table, but also the presence of our
audit field trigger.  That check should be an option check for base tables
following our regime, not a requirement.

Note that this fix not only allows us to get view data models, but also non-
utils linked tables.