Refactor item wrkbnch initialization sequence
When itemAvailabilityWorkbench runs, it doesn't initialize the embedded "item"
form properly by calling "set" on the item form.  Also, other item form signals
don't seem to get called until after the first item has been loaded via
item::setId(int).  So, we need to initialize the item scripts, but we cannot
do this for every time the item id changes itemAvailabilityWorkbench... there
are better signals to latch onto from the item form itself (newId/populated)
and for some reason after the first time we cannot re-acquire a reference to the
embedded item form from the availability workbench: mywindow.findChild("item")
returns null at some time after the first time it's found... that's got to be a
bug of some sort, but we'll work around that for now.

Note that this is a breaking change as there are now different expectations of
item scripts as far as responding to changes in the item id both from the
availability workbench and the item form itself (though an id change directly
there is not possible.)

This is in part addressing:
https://gitlab.com/musesystems-od/support/issues/69

Note that updated fds and musesuperchar packages will be required.