Refactor Config JS API to return JS typed values
In the JavaScript side of the house, we really want to work with more JSish
types and constructs.  So, if we have an array, we want a JavaScript array, not
the text representation of the PostgreSQL array; if we have a PG JSONB value,
we want a JS Object Literal.