Refactor formatting (?)
Some Sublime Text 3 plugin (probably prettier) has decided to reformat markdown
files.  While I can't see a way to turn this off immediately, it doesn't appear
to be hurting anything so lets let it go for now... the downsides to not
automatically formatting with prettier are less than the downsides of casting
this wide of a net for formatting.